use bcl healthcheck
