--- app/pages/health/status.js
@@ -1,14 +1,0 @@
-export const status = {
-  healthy: {
-    code: 200,
-    message: 'Healthy',
-  },
-  degraded: {
-    code: 200,
-    message: 'Degraded',
-  },
-  unhealthy: {
-    code: 503,
-    message: 'Unhealthy',
-  },
-};
