+++ package.json
@@ -37,7 +37,7 @@
     "axios": "^0.19.0",
     "body-parser": "^1.19.0",
     "buying-catalogue-components": "^1.1.56",
-    "buying-catalogue-library": "^1.0.5",
+    "buying-catalogue-library": "^1.0.11",
     "compression": "^1.7.4",
     "connect-redis": "^4.0.4",
     "connect-slashes": "^1.4.0",
